LayoutTests/imported/w3c/web-platform-tests/trusted-types/tt-block-eval-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS eval blocks if the default policy rejects a value.
